viki-ai/medication_extraction/pyproject.toml
[project]
name = "medication-extraction"
version = "0.1.0"
description = "Medication extraction service using FastAPI and Vertex AI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "google-cloud-aiplatform>=1.38.1",
    "vertexai>=0.0.1",
    "pypdf>=5.1.0",
    "gcloud-aio-storage>=9.3.0",
    "google-auth>=2.37.0",
    "google-cloud-pubsub>=2.27.2",
    "google-cloud-documentai>=3.1.0",
    "google-cloud-tasks>=2.18.0",
    "google-cloud-logging>=3.11.3",
    "google-cloud-firestore>=2.19.0",
    "coloredlogs>=15.0.1",
    "psycopg2==2.9.10",
    "psycopg2-binary==2.9.10",
    "google-genai>=1.29.0",
    "setuptools<81",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
pythonpath = ["src"]
